// src/application/dtos/user/userCreate.dto.ts
export class UserCreateDto {
  address: string = "";
  age: string = "";
  birthday: string = "";
  cellPhone: string = "";
  documentNumber: string = "";
  email: string = "";
  idDocumentsType: number = 0;
  idRole: number = 0;
  lastName: string = "";
  names: string = "";
  password: string = "";
  repeatPassword: string = "";
  state: string = "";
  userCreate: number = 0;
}